tangle - extract common backups (#141154)
